+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [4.22.0][] - 2026-09-03
+
 ### Changed
 
 -   `@lumx/react`, `@lumx/vue`:
@@ -2992,7 +2994,8 @@ _Failed released_
 [3.7.0]: https://github.com/lumapps/design-system/compare/v3.6.8...v3.7.0
 [3.6.8]: https://github.com/lumapps/design-system/compare/v3.6.7...v3.6.8
 [3.6.7]: https://github.com/lumapps/design-system/tree/v3.6.7
-[Unreleased]: https://github.com/lumapps/design-system/compare/4.21.0...HEAD
+[Unreleased]: https://github.com/lumapps/design-system/compare/4.22.0...HEAD
+[4.22.0]: https://github.com/lumapps/design-system/compare/4.21.0...4.22.0
 [4.21.0]: https://github.com/lumapps/design-system/compare/4.20.0...4.21.0
 [4.20.0]: https://github.com/lumapps/design-system/compare/4.19.0...4.20.0
 [4.19.0]: https://github.com/lumapps/design-system/compare/4.18.0...4.19.0